Move Run function to run.go
And move logging setup to logging.go
